+++ src/gitmojiCommit/composeCommit.js
@@ -1,13 +1,6 @@
-const { workspace } = require('vscode')
-const convertToEmoji = require('./convertToEmoji')
+const convertToEmoji = require('./convertToEmoji');
 
 module.exports = (commitType, commitText) => {
-  const multiline = workspace.getConfiguration().get('gitmoji.multilineCommit')
-  const emoji = convertToEmoji(commitType)
-
-  if (multiline) {
-    return `git commit -m $'${emoji} ${commitText}'`
-  } else {
-    return `git commit -m "${emoji} ${commitText}"`
-  }
-}
+  const emoji = convertToEmoji(commitType);
+  return `${emoji} ${commitText}`;
+};
