--- src/gitmojiCommit/doesTerminalExist.js
@@ -1,5 +1,0 @@
-const { window } = require('vscode')
-
-module.exports = () => {
-  return window.terminals.length !== 0
-}
